[test] disable splash screen
To see whether it is the cause of https://github.com/musescore/MuseScore/issues/23970. The reason for conjecturing that, is that this Qt bug https://bugreports.qt.io/browse/QTBUG-124750 has the same backtrace and mentions closing a widget; the only widget being closed during startup is the splash screen.